ComandesActualitzadesProfileFet

+++ database/migrations/2024_11_26_000020_create_producte_table.php
@@ -6,9 +6,6 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::create('producte', function (Blueprint $table) {
@@ -16,16 +13,13 @@ return new class extends Migration
             $table->string('nom');
             $table->string('descripcio')->nullable();
             $table->string('imatge')->nullable();
-            $table->integer('preu');
+            $table->decimal('preu', 10, 2);
             $table->foreignId('categoria_id')->constrained(table:'categoria')->onUpdate('cascade')->onDelete('cascade');
             $table->timestamp('created_at')->useCurrent();
             $table->timestamp('updated_at')->useCurrent();
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('producte');
